Lots of small things.
Created a script that generates a small subset of the data,
modified file_loader to 1. Not load data it's not using (warning: hacky, all I did was comment things out), 2. accept a 'tiny' parameter to only load the small subset
Modified 'main_tweak' to work on the tiny dataset (so I can play with architecture stuff without running out of RAM), and also annotated the modeler.stdn parameters (as identified from the input data) so we don't have to generate it from the input data in the future.
WIP load_example.py script to show how to load a model from memory.

Current future goals: Understand data format, make it possible to load from a file, and then clean up the new code. Possibly change the architecture such that data-sampling is decoupled from model creation.